backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic-settings==2.2.1
grpcio==1.63.0
protobuf==5.26.1
PyJWT==2.8.0
openai==1.30.0
tenacity==8.3.0
httpx==0.27.0
uniphore-protos @ git+https://github.com/uniphore/protos-python.git
